requirements.txt
Django==3.1.2
django-crispy-forms==1.9.0
django-filter==2.4.0
django-summernote==0.8.11.6
djangorestframework==3.11.0
Pillow>=7.0.0
PyPDF2==1.26.0
xhtml2pdf==0.2.5
gunicorn==20.0.4
